# api/openapi.yaml
openapi: 3.0.3
info:
  title: GHTA API
  version: "1.0"
  description: >
    Multi-source trend aggregation API. GitHub is the first source; App Store,
    Chrome Web Store and Microsoft Store adapters conform to the same model.
    This contract is frozen for the frontend (change 5).
servers:
  - url: http://localhost:3000
paths:
  /health:
    get:
      summary: Liveness probe
      responses:
        "200": { description: OK }

  /trending:
    get:
      summary: List tracked items with filtering and sorting
      parameters:
        - { name: source, in: query, schema: { type: string, enum: [github, appstore, chrome, msstore] }, description: "Filter by source; omit for all sources" }
        - { name: stars, in: query, schema: { type: string }, description: "Range on metrics.stars: a..b, >n, <n, n (0 is valid)" }
        - { name: issues, in: query, schema: { type: string }, description: "Range on metrics.openIssues" }
        - { name: language, in: query, schema: { type: string } }
        - { name: category, in: query, schema: { type: string }, description: "categoryId (hex) or a category path containing '/' (e.g. infra/containers)" }
        - { name: type, in: query, schema: { type: string, enum: [cli, app, library, software, tutorial, awesome, interview, skill] }, description: "Form facet filter" }
        - { name: topics, in: query, schema: { type: string }, description: "Comma-separated topics; every one must be present" }
        - { name: sort, in: query, schema: { type: string }, description: "field:order; field in [stars,forks,issues,fetchedAt] (stars aliases metrics.stars)" }
        - { name: limit, in: query, schema: { type: integer, maximum: 50, default: 50 } }
      responses:
        "200":
          description: Matching items
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/TrackedItem" }
        "400": { description: Invalid filter/sort/limit }

  /trending/rising:
    get:
      summary: Items ranked by growth in a window
      parameters:
        - { name: window, in: query, schema: { type: string, enum: [daily, weekly, monthly], default: weekly } }
        - { name: source, in: query, schema: { type: string } }
        - { name: category, in: query, schema: { type: string } }
        - { name: language, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, maximum: 50, default: 50 } }
      responses:
        "200":
          description: Items with a non-null increase for the window, highest first
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/TrackedItem" }
        "400": { description: Invalid window/limit }

  /trending/item:
    get:
      summary: One item plus its snapshot history
      parameters:
        - { name: source, in: query, required: true, schema: { type: string } }
        - { name: externalId, in: query, required: true, schema: { type: string } }
      responses:
        "200":
          description: Item and recent metric snapshots
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      item: { $ref: "#/components/schemas/TrackedItem" }
                      history:
                        type: array
                        items: { $ref: "#/components/schemas/MetricSnapshot" }
        "400": { description: Missing source/externalId }
        "404": { description: Not found }

  /category/facets:
    get:
      summary: Type-facet enum with item counts (public, for filter chips)
      responses:
        "200":
          description: Form types in priority order, each with its item count
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      type:
                        type: array
                        items:
                          type: object
                          properties:
                            key: { type: string }
                            name: { type: string }
                            count: { type: integer }

  /category:
    get:
      summary: Domain category tree with item counts (public, createdBy=taxonomy only)
      responses:
        "200":
          description: Nested category tree; each node carries an item count (distinct items for parents)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/CategoryTree" }
    post:
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CategoryInput" }
      responses:
        "201": { description: Created }
        "400": { description: Invalid body }
  /category/{id}:
    get:
      summary: Get a category
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: OK }
        "404": { description: Not found }
    patch:
      summary: Update a category
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CategoryInput" }
      responses:
        "200": { description: Updated }
        "404": { description: Not found }
    delete:
      summary: Delete a category
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: Deleted }
        "404": { description: Not found }

  /user:
    get:
      summary: List users
      responses:
        "200": { description: OK }
    post:
      summary: Create a user
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserInput" }
      responses:
        "201": { description: Created }
  /user/{id}:
    get:
      summary: Get a user
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: OK }
        "404": { description: Not found }
    patch:
      summary: Update a user
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserInput" }
      responses:
        "200": { description: Updated }
        "404": { description: Not found }
    delete:
      summary: Delete a user
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: Deleted }
        "404": { description: Not found }

components:
  schemas:
    TrackedItem:
      type: object
      properties:
        id: { type: string }
        source: { type: string }
        externalId: { type: string }
        name: { type: string }
        description: { type: string }
        language: { type: string }
        categoryId: { type: array, items: { type: string } }
        categoryPath: { type: array, items: { type: string } }
        type: { type: string, description: form facet (cli|app|library|software|tutorial|awesome|interview|skill) }
        primaryMetric: { type: string }
        metricDirection: { type: string, enum: [desc-better, asc-better] }
        metrics: { type: object, additionalProperties: { type: number } }
        dailyIncrease: { type: number, nullable: true }
        weeklyIncrease: { type: number, nullable: true }
        monthlyIncrease: { type: number, nullable: true }
        analysisStatus: { type: string, enum: [pending, done, failed] }
        sourceData: { type: object, additionalProperties: true }
        fetchedAt: { type: string, format: date-time }
    MetricSnapshot:
      type: object
      properties:
        capturedAt: { type: string, format: date-time }
        metrics: { type: object, additionalProperties: { type: number } }
    CategoryTree:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        path: { type: string }
        count: { type: integer, description: "Items filed under this node (distinct items for a parent)" }
        children:
          type: array
          items: { $ref: "#/components/schemas/CategoryTree" }
    CategoryInput:
      type: object
      required: [name, path]
      properties:
        name: { type: string }
        description: { type: string }
        parentId: { type: string, nullable: true }
        path: { type: string }
    UserInput:
      type: object
      properties:
        email: { type: string }
        name: { type: string }
        role: { type: string, enum: [admin, user] }